backend/requirements.txt
cryptography==38.0.1
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
music-tag==0.4.3
Pillow==9.2.0
PyJWT==2.4.0
requests==2.28.1
django-cors-headers==3.13.0
